moodboard-project/worker/requirements.txt
redis==4.5.5
rq==1.15.1
python-dotenv==1.0.0
Pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9